[#7] feat: check nickname

+++ server/src/main/java/com/gongsp/api/service/UserService.java
@@ -3,5 +3,5 @@ package com.gongsp.api.service;
 import com.gongsp.db.entity.User;
 
 public interface UserService {
-    User getUserByUserSeq(Long userSeq);
+    User getUserByUserSeq(Integer userSeq);
 }
